Fix bug associated with small patch stride
Fixed bug in which, at small values of patch stride, output segmentation map did not seem to match likelihood map. Modified manner in which pixels are categorized as brain. For each pixel, the number of times inference looked at that pixel is considered. The percentage of times inference categorized that pixel as brain is converted to a float, then compared to the selected threshold. In most instances this produces very similar, if not identical, results as before, but is robust to situations in which each pixel has inference run on it very many times. Provided option to switch back to old calculation method should it be desirable.