Add anomalous mask area detection
Added detection of anomalous mask area - low. If the percentage of an image taken up by a slice's mask is less than both of its neighbors, it will be flagged in quality check, column 'notes_2'. If the same slice is flagged by both this method and standard quality check model, the entry in quality_check.csv will not be duplicated, but both notes columns in the file will be non-empty.
Resolved an indentation issue in quality.py