Add likelihood map creation option
Added option to switch between two methods of creating the likelihood map that is binarized to create output mask. Switch is likelihood categorization, -lc True/False. If True, we calculate each pixel of the likelihood map as the mean value of the likelihood values for each patch containing that pixel. If False, we instead take the maximum value.
-lc True is recommended, as using the maximum value tends to incorrectly categorize a relatively large number of pixels as brain. As such, True is the default value