Improve Exception Handling
Added more graceful handling of exceptions during segmentation, particulary for 'dataset' and 'directory' input types. If there is an exception during segmentation for an image, that segmentation will fail, the details of the exception along with the traceback will be placed in segmentation_log.csv, and the program will move to the next image to segment.

Added more graceful handling of exceptions during quality checking for all input types. Feature calculation errors, including tracebacks, are now included in segmentation_log.txt. File and slice are also specified.

Now allow for additional files (with extensions not .nii or .nii.gz) to be present in mouse and modality directories in 'dataset' input type, and in directory in 'directory' input type. For 'dataset' input type, we expect that each mouse/modality directory has exactly one .nii file, and raise an exception if there are zero or more than one. For 'directory', we assume that each .nii or .nii.gz file in the specified directory should have inference run on it. If there are zero files with the relevant extensions, raise an exception.