Automatic Orientation Detection
Added capacity for automatic detection of input orientation. If input argument --ouput_orientation (-oo) is set to 'auto', all files will be written in the same orientation as the output. If -oo is set to an acceptable NIfTI orientation string (three characters), then the output will be written in that orientation.