Implement reorientation solution
Implemented two state reorientation solution.
To ensure that output masks have an orientation that matches the orientation of the input data, we copy relevant metadata information from the input to the mask. Relevant orientation-defining metadata includes direction, origin, and spacing.
To ensure that the data on which segmentation is performed matches the orientation of images on which the neural network was trained, we define use a two step process requiring two pieces of user input: input MRI plane, whether it is necessary to flip the resulting images 180 degrees. With knowledge of input MRI plane, we can re-slice images to the axial plane on which the CNN is trained. It is up to the user to identify if the resulting axial image must be flipped 180 degrees such that the top of the brain is pointing downwards.
Few additional bugfixes, included mis-spelling 'sagittal' as 'saggital.'